langchain-demo/pyproject.toml
[project]
name = "langchain-demo"
version = "0.1.0"
description = "My langchain learning demo."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "agentrun-sdk[playwright,server]>=0.0.8",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.5",
    "langgraph>=1.0.6",
    "mss>=10.1.0",
    "playwright>=1.40.0",
    "pyautogui>=0.9.54",
    "pycodestyle>=2.14.0",
    "python-dotenv>=1.0.0",
    "rich>=14.2.0",
]
